Move server mock to own package
